+++ packages/@ember/-internals/glimmer/lib/components/input.ts
@@ -4,7 +4,7 @@
 import hasDOM from '@ember/-internals/browser-environment/lib/has-dom';
 import { type Opaque } from '@ember/-internals/utility-types';
 import { assert, warn } from '@ember/debug';
-import { action } from '@ember/object';
+import { action } from '@ember/object/action';
 import { valueForRef } from '@glimmer/reference/lib/reference';
 import { untrack } from '@glimmer/validator/lib/tracking';
 import InputTemplate from '../templates/input';
